Rename backend to apiserver
This PR renames the vmclarity-backend cmd and backend packages to
vmclarity-apiserver and apiserver respectively in preparation to split
the orchestrator and uibackend out of the backend monolith.
