+++ pkg/apiserver/database/gorm/common.go
@@ -23,7 +23,7 @@ import (
 	jsonpatch "github.com/evanphx/json-patch"
 	"gorm.io/gorm"
 
-	"github.com/openclarity/vmclarity/pkg/backend/database/types"
+	"github.com/openclarity/vmclarity/pkg/apiserver/database/types"
 	"github.com/openclarity/vmclarity/pkg/shared/utils"
 )
