+++ pkg/apiserver/rest/server.go
@@ -25,8 +25,8 @@ import (
 	echomiddleware "github.com/labstack/echo/v4/middleware"
 
 	"github.com/openclarity/vmclarity/api/server"
-	"github.com/openclarity/vmclarity/pkg/backend/common"
-	databaseTypes "github.com/openclarity/vmclarity/pkg/backend/database/types"
+	"github.com/openclarity/vmclarity/pkg/apiserver/common"
+	databaseTypes "github.com/openclarity/vmclarity/pkg/apiserver/database/types"
 	"github.com/openclarity/vmclarity/pkg/shared/log"
 	uiserver "github.com/openclarity/vmclarity/pkg/uibackend/api/server"
 	uirest "github.com/openclarity/vmclarity/pkg/uibackend/rest"
